EntityFrameworkQuery refactored, partially renamed to EntityFrameworkPostProcessingQuery.
