+++ src/Infrastructure/Riganti.Utils.Infrastructure.EntityFramework/EntityFrameworkQuery.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Data.Entity;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Riganti.Utils.Infrastructure.Core;
 
 namespace Riganti.Utils.Infrastructure.EntityFramework
@@ -11,12 +7,13 @@ namespace Riganti.Utils.Infrastructure.EntityFramework
     /// <summary>
     /// A base implementation of query object in Entity Framework.
     /// </summary>
-    public abstract class EntityFrameworkQuery<TResult> : EntityFrameworkQuery<TResult, TResult>
+    public abstract class EntityFrameworkQuery<TResult> : EntityFrameworkPostProcessingQuery<TResult, TResult>
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="EntityFrameworkQuery{TResult}"/> class.
         /// </summary>
-        protected EntityFrameworkQuery(IUnitOfWorkProvider unitOfWorkProvider) : base(unitOfWorkProvider)
+        protected EntityFrameworkQuery(IUnitOfWorkProvider unitOfWorkProvider)
+            : base(unitOfWorkProvider)
         {
         }
 
@@ -33,62 +30,24 @@ namespace Riganti.Utils.Infrastructure.EntityFramework
     /// <summary>
     /// A base implementation of query object in Entity Framework.
     /// </summary>
-    public abstract class EntityFrameworkQuery<TQueryableResult, TResult> : EntityFrameworkQuery<TQueryableResult, TResult, DbContext>
-    {
-        protected EntityFrameworkQuery(IUnitOfWorkProvider unitOfWorkProvider) : base(unitOfWorkProvider)
-        {
-        }
-    }
-
-
-    /// <summary>
-    /// A base implementation of query object in Entity Framework.
-    /// </summary>
-    public abstract class EntityFrameworkQuery<TQueryableResult, TResult, TDbContext> : QueryBase<TQueryableResult, TResult>
+    public abstract class EntityFrameworkQuery<TResult, TDbContext> : EntityFrameworkPostProcessingQuery<TResult, TResult, TDbContext>
         where TDbContext : DbContext
     {
-        private readonly IUnitOfWorkProvider unitOfWorkProvider;
-
         /// <summary>
-        /// Initializes a new instance of the <see cref="EntityFrameworkQuery{TQueryableResult, TResult, TDbContext}"/> class.
-        /// </summary>
-        protected EntityFrameworkQuery(IEntityFrameworkUnitOfWorkProvider<TDbContext> unitOfWorkProvider)
-            : this((IUnitOfWorkProvider)unitOfWorkProvider)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="EntityFrameworkQuery{TQueryableResult, TResult, TDbContext}"/> class.
+        /// Initializes a new instance of the <see cref="EntityFrameworkQuery{TResult}"/> class.
         /// </summary>
         protected EntityFrameworkQuery(IUnitOfWorkProvider unitOfWorkProvider)
+            : base(unitOfWorkProvider)
         {
-            this.unitOfWorkProvider = unitOfWorkProvider;
         }
 
         /// <summary>
-        /// Gets the <see cref="DbContext"/>.
+        ///     When overriden in derived class, it allows to modify the materialized results of the query before they are returned
+        ///     to the caller.
         /// </summary>
-        protected virtual TDbContext Context
+        protected override IList<TResult> PostProcessResults(IList<TResult> results)
         {
-            get
-            {
-                var context = EntityFrameworkUnitOfWork.TryGetDbContext<TDbContext>(unitOfWorkProvider);
-                if (context == null)
-                {
-                    throw new InvalidOperationException("The EntityFrameworkQuery must be used in a unit of work of type EntityFrameworkUnitOfWork!");
-                }
-                return context;
-            }
-        }
-
-        protected override async Task<IList<TQueryableResult>> ExecuteQueryAsync(IQueryable<TQueryableResult> query, CancellationToken cancellationToken)
-        {
-            return await query.ToListAsync(cancellationToken);
-        }
-
-        public override async Task<int> GetTotalRowCountAsync(CancellationToken cancellationToken)
-        {
-            return await GetQueryable().CountAsync(cancellationToken);
+            return results;
         }
     }
 }
